# No more boomerangs