@mapsam 2018-01-24
A few improvements in https://github.com/mapbox/vtquery/pull/53/commits/b1887e0244437bddc581270599843f3c0bed7a86 and some quick 👀  and help from @springmeyer over in https://github.com/mapbox/vtquery/pull/53/commits/79ebff1c938eee9cb9785b12c74a092865cdc47e seem to have improved things a bit!

[after] https://github.com/mapbox/vtquery/pull/53/commits/79ebff1c938eee9cb9785b12c74a092865cdc47e

```
1. pip: many building polygons ... 969 runs/s (1032ms)
2. pip: many building polygons, single layer ... 1044 runs/s (958ms)
3. query: many building polygons, single layer ... 805 runs/s (1242ms)
4. query: linestrings, mapbox streets roads ... 1582 runs/s (632ms)
5. query: polygons, mapbox streets buildings ... 804 runs/s (1244ms)
6. query: all things - dense single tile ... 434 runs/s (2303ms)
7. query: all things - dense nine tiles ... 51 runs/s (19684ms)
8. elevation: terrain tile nepal ... 845 runs/s (1184ms)
9. geometry: 2000 points in a single tile, no properties ... 1980 runs/s (505ms)
10. geometry: 2000 points in a single tile, with properties ... 1277 runs/s (783ms)
11. geometry: 2000 linestrings in a single tile, no properties ... 1100 runs/s (909ms)
12. geometry: 2000 linestrings in a single tile, with properties ... 723 runs/s (1384ms)
13. geometry: 2000 polygons in a single tile, no properties ... 684 runs/s (1461ms)
14. geometry: 2000 polygons in a single tile, with properties ... 524 runs/s (1908ms)

```