@mapsam 2018-01-24
I'm feeling really good about this update - there may be a small decrease in performance for queries, but the improved readability leads to easier maintenance, plus this improves point in polygon queries. Going to merge into the `dedupe` branch!